Removed unnecessary log info and added an index to the output file
In case of RNAseq (tophat) all files are named “accepted_hits.bam“.
Current version uses the bam name to export to a single folder and
woful thereby overwrite the previous file after every iteration.
Temporarily solution is the addition of the index as a prefix to the
output name. A better solution will have to be incorporated.