Removed mapping quality from score calculation
The bug previously discovered, which led to a incorrect ‘edit distance’
score from the human read (reverse) in PE sequencing, shed new light on
the usefulness of the mapping quality in the calculations of the
‘score’.
Better results were found when only taking the edit distance into
account. Therefore the mapping quality was taken out of the equation.