Strict filtering only - filter step used in manuscript
Multiple paper filter based on the simple rule ‘remove everything that
maps to the mouse reference’. This version of XenofilteR does not check
for lower edit distance or any of the other features in the normal
version. This version just filters all reads from the human bam file
that also map to the mouse reference genome.
This version is not for general use, only as a comparison in the
manuscript.